requirements.txt
torch
torchvision
Ninja
gradio
huggingface_hub
hf_transfer
Pillow==9.5.0
psutil
imageio
scikit-image
lpips
wandb
dlib